move HiltUtils to package utils
