+++ app/src/test/java/com/github/sdpsharelook/utils/HiltUtils.kt
@@ -1,4 +1,4 @@
-package com.github.sdpsharelook
+package com.github.sdpsharelook.utils
 
 import android.content.ComponentName
 import android.content.Intent
